API+Metiers+CRUD
CRUD + API mailing +API notification
 +API weather data
