+++ Gestion-User/services/APIConnector.java
@@ -1,0 +1,90 @@
+package edu.SprintJava2.services;
+
+import org.json.simple.JSONArray;
+import org.json.simple.JSONObject;
+import org.json.simple.parser.JSONParser;
+
+import java.net.HttpURLConnection;
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.Scanner;
+
+
+public class APIConnector {
+
+    private final String urlString;
+
+
+    public APIConnector(String urlString) throws MalformedURLException {
+        this.urlString = urlString;
+    }
+
+    public JSONArray getJSONArray(String query){
+        try {
+            URL url = new URL(urlString + query);
+
+            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+            conn.setRequestMethod("GET");
+            conn.connect();
+
+            //Check if connect is made
+            int responseCode = conn.getResponseCode();
+
+            if (responseCode != 200) {
+                throw new RuntimeException("HttpResponseCode: " + responseCode);
+            } else {
+
+                StringBuilder informationString = new StringBuilder();
+                Scanner scanner = new Scanner(url.openStream());
+
+                while (scanner.hasNext()) {
+                    informationString.append(scanner.nextLine());
+                }
+                scanner.close();
+
+                JSONParser parse = new JSONParser();
+                
+                System.out.println("edu.SprintJava2.services.APIConnector.getJSONArray() : ");
+                System.out.println(informationString);
+
+                return (JSONArray) parse.parse(String.valueOf(informationString));
+            }
+        }catch (Exception e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    public JSONObject getJSONObject(String query){
+        try {
+            URL url = new URL(urlString + query);
+
+            HttpURLConnection conn = (HttpURLConnection) url.openConnection();
+            conn.setRequestMethod("GET");
+            conn.connect();
+
+            //Check if connect is made
+            int responseCode = conn.getResponseCode();
+
+            if (responseCode != 200) {
+                throw new RuntimeException("HttpResponseCode: " + responseCode);
+            } else {
+
+                StringBuilder informationString = new StringBuilder();
+                Scanner scanner = new Scanner(url.openStream());
+
+                while (scanner.hasNext()) {
+                    informationString.append(scanner.nextLine());
+                }
+                scanner.close();
+
+                JSONParser parse = new JSONParser();
+
+                return (JSONObject) parse.parse(String.valueOf(informationString));
+            }
+        }catch (Exception e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+}
