+++ Gestion-User/views/ContactPage.fxml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.ContactPageController">
+   <children>
+      <AnchorPane prefHeight="107.0" prefWidth="1043.0" style="-fx-background-color: ADC47F;">
+         <children>
+            <Label layoutX="373.0" layoutY="27.0" prefHeight="21.0" prefWidth="389.0" text="Contact Us " textFill="WHITE">
+               <font>
+                  <Font name="System Bold" size="37.0" />
+               </font>
+            </Label>
+            <Button layoutX="940.0" layoutY="27.0" mnemonicParsing="false" onAction="#retourner" text="Retourner" />
+         </children>
+      </AnchorPane>
+      <Label layoutX="167.0" layoutY="185.0" prefHeight="32.0" prefWidth="143.0" text="Name" />
+      <Label layoutX="164.0" layoutY="267.0" prefHeight="32.0" prefWidth="93.0" text="Email" />
+      <Label layoutX="165.0" layoutY="347.0" prefHeight="21.0" prefWidth="47.0" text="Date" />
+      <Label layoutX="163.0" layoutY="425.0" text="Description" />
+      <TextField fx:id="colname" layoutX="164.0" layoutY="225.0" />
+      <TextField fx:id="colemail" layoutX="163.0" layoutY="300.0" />
+      <TextField fx:id="coldesc" layoutX="163.0" layoutY="455.0" prefHeight="158.0" prefWidth="307.0" />
+      <TextField fx:id="coldate" layoutX="163.0" layoutY="375.0" />
+      <Button layoutX="583.0" layoutY="534.0" mnemonicParsing="false" onAction="#envoyercontact" text="Envoyer" />
+      <Button layoutX="693.0" layoutY="534.0" mnemonicParsing="false" onAction="#supprimercontact" text="Supprimer" />
+   </children>
+</AnchorPane>
