+++ Gestion-User/views/InterfaceAdmin.fxml
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.layout.VBox?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane id="AnchorPane" maxHeight="630.0" maxWidth="1043.0" minHeight="630.0" minWidth="1043.0" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.InterfaceAdminController">
+   <children>
+      <AnchorPane layoutY="-6.0" prefHeight="636.0" prefWidth="200.0" style="-fx-background-color: ADC47F;">
+         <children>
+            <Button fx:id="userpagebtn" layoutX="33.0" layoutY="94.0" mnemonicParsing="false" onAction="#userpage" prefHeight="31.0" prefWidth="116.0" text="Users" />
+         </children>
+      </AnchorPane>
+      <AnchorPane fx:id="adminpage" layoutX="200.0" prefHeight="106.0" prefWidth="842.0">
+         <children>
+            <Label fx:id="pagelabel" layoutX="33.0" layoutY="9.0" prefHeight="88.0" prefWidth="776.0" textFill="#c93d3d" />
+            <Button fx:id="logoutbtn" layoutX="744.0" layoutY="22.0" mnemonicParsing="false" onAction="#logout" text="log out" />
+            <Label layoutX="24.0" layoutY="38.0" prefHeight="45.0" prefWidth="287.0" text="Welcome Admin">
+               <font>
+                  <Font size="30.0" />
+               </font>
+            </Label>
+            <ImageView fitHeight="94.0" fitWidth="100.0" layoutX="264.0" layoutY="14.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../../../../../../Downloads/admin.png" />
+               </image>
+            </ImageView>
+         </children>
+      </AnchorPane>
+      <VBox fx:id="vbox" layoutX="220.0" layoutY="128.0" prefHeight="477.0" prefWidth="786.0" />
+   </children>
+</AnchorPane>
