+++ Gestion-User/views/InterfaceLogin.fxml
@@ -1,0 +1,115 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.CheckBox?>
+<?import javafx.scene.control.ComboBox?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.paint.RadialGradient?>
+<?import javafx.scene.paint.Stop?>
+<?import javafx.scene.shape.Rectangle?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="969.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.InterfaceLoginController">
+   <children>
+      <AnchorPane layoutY="-3.0" prefHeight="115.0" prefWidth="1043.0" style="-fx-background-color: ADC47F;">
+         <children>
+            <ImageView fitHeight="124.0" fitWidth="171.0" layoutX="659.0" layoutY="11.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../images/Image2.png" />
+               </image>
+            </ImageView>
+            <Label layoutX="141.0" layoutY="44.0" prefHeight="45.0" prefWidth="535.0" text="Welcome to our platform Nature Cruise" textFill="WHITE">
+               <font>
+                  <Font name="Gabriola" size="40.0" />
+               </font>
+            </Label>
+            <Label layoutX="679.0" layoutY="93.0" text="Camping made easy " />
+            <ImageView fitHeight="39.0" fitWidth="40.0" layoutX="984.0" layoutY="14.0" onMouseClicked="#closeapplication" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="file:/C:/Users/chayma/Downloads/close.png" />
+               </image>
+            </ImageView>
+         </children>
+      </AnchorPane>
+      <AnchorPane layoutX="22.0" layoutY="163.0" prefHeight="467.0" prefWidth="417.0">
+         <children>
+            <Label layoutX="20.0" layoutY="64.0" text="Name" />
+            <TextField fx:id="TFname" layoutX="20.0" layoutY="91.0" prefHeight="31.0" prefWidth="139.0" promptText="name" />
+            <Label layoutX="172.0" layoutY="64.0" text="Lastname" />
+            <TextField fx:id="TFlastname" layoutX="172.0" layoutY="91.0" prefHeight="31.0" prefWidth="139.0" promptText="lastname" />
+            <Label layoutX="22.0" layoutY="147.0" text="Email" />
+            <TextField fx:id="TFemail" layoutX="20.0" layoutY="174.0" prefHeight="31.0" prefWidth="295.0" promptText="email" />
+            <Label layoutX="21.0" layoutY="223.0" text="Password" />
+            <PasswordField fx:id="TFpassword" layoutX="20.0" layoutY="252.0" prefHeight="31.0" prefWidth="295.0" promptText="***********" />
+            <Label layoutX="103.0" layoutY="-10.0" text="Sign Up" textFill="#a6bc7c">
+               <font>
+                  <Font name="System Bold" size="27.0" />
+               </font>
+            </Label>
+            <ImageView fitHeight="54.0" fitWidth="90.0" layoutX="226.0" layoutY="-17.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../images/check.png" />
+               </image>
+            </ImageView>
+            <ImageView fx:id="sigupimage" fitHeight="68.0" fitWidth="82.0" layoutX="356.0" layoutY="325.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../images/userlogo.png" />
+               </image></ImageView>
+            <Button layoutX="345.0" layoutY="407.0" mnemonicParsing="false" onAction="#uploadsiguppic" text="upload pic" />
+            <Label layoutX="22.0" layoutY="304.0" text="Role" />
+            <Button layoutX="119.0" layoutY="407.0" mnemonicParsing="false" onAction="#ajouteruser" prefHeight="31.0" prefWidth="139.0" text="Sign Up" />
+            <ComboBox fx:id="combobox1" layoutX="21.0" layoutY="336.0" prefWidth="150.0" />
+         </children>
+      </AnchorPane>
+      <AnchorPane layoutX="585.0" layoutY="141.0" prefHeight="403.0" prefWidth="399.0">
+         <children>
+            <Label layoutX="36.0" layoutY="74.0" prefHeight="21.0" prefWidth="54.0" text="Email" />
+            <TextField fx:id="TFemail2" layoutX="36.0" layoutY="105.0" prefHeight="31.0" prefWidth="281.0" promptText="Your email" />
+            <PasswordField fx:id="TFpassword2" layoutX="38.0" layoutY="207.0" prefHeight="31.0" prefWidth="281.0" promptText="**********" />
+            <CheckBox fx:id="cbox" layoutX="238.0" layoutY="252.0" mnemonicParsing="false" onAction="#checkbox" text="View Password" />
+            <Label layoutX="36.0" layoutY="164.0" prefHeight="21.0" prefWidth="95.0" text="Password" />
+            <Button layoutX="251.0" layoutY="296.0" mnemonicParsing="false" onAction="#validate" text="Sign In" />
+            <Label layoutX="119.0" layoutY="12.0" prefHeight="21.0" prefWidth="109.0" text="Sign In" textFill="#8eb970">
+               <font>
+                  <Font name="Arial Rounded MT Bold" size="28.0" />
+               </font>
+            </Label>
+            <ImageView fitHeight="61.0" fitWidth="81.0" layoutX="235.0" layoutY="6.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="file:/C:/Users/chayma/Downloads/sign-in.png" />
+               </image>
+            </ImageView>
+            <ImageView fitHeight="31.0" fitWidth="30.0" layoutX="-1.0" layoutY="106.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="file:/C:/Users/chayma/Downloads/user.png" />
+               </image>
+            </ImageView>
+            <ImageView fitHeight="30.0" fitWidth="30.0" layoutX="-1.0" layoutY="208.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="file:/C:/Users/chayma/Downloads/key.png" />
+               </image>
+            </ImageView>
+            <ImageView fitHeight="30.0" fitWidth="30.0" layoutX="377.0" layoutY="248.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="file:/C:/Users/chayma/Downloads/eye.png" />
+               </image>
+            </ImageView>
+         </children>
+      </AnchorPane>
+      <Rectangle arcHeight="5.0" arcWidth="5.0" height="318.0" layoutX="522.0" layoutY="190.0" stroke="BLACK" strokeType="INSIDE" style="-fx-arc-height: ADC47F;" width="14.0">
+         <fill>
+            <RadialGradient centerX="0.5" centerY="0.5" radius="0.5">
+               <stops>
+                  <Stop color="#aadd9f" />
+                  <Stop color="WHITE" offset="1.0" />
+               </stops>
+            </RadialGradient>
+         </fill>
+      </Rectangle>
+   </children>
+</AnchorPane>
