+++ Gestion-User/views/Userprofile.fxml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.PasswordField?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.image.Image?>
+<?import javafx.scene.image.ImageView?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.shape.Circle?>
+<?import javafx.scene.text.Font?>
+<?import javafx.scene.text.Text?>
+
+<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.UserprofileController">
+   <children>
+      <AnchorPane layoutX="-7.0" prefHeight="173.0" prefWidth="1063.0" style="-fx-background-color: ADC47F;">
+         <children>
+            <Circle fx:id="profilepicture" fill="WHITE" layoutX="128.0" layoutY="87.0" radius="70.0" stroke="BLACK" strokeType="INSIDE" />
+            <Label fx:id="namellastnameuser" layoutX="208.0" layoutY="25.0" prefHeight="41.0" prefWidth="258.0" />
+            <ImageView fitHeight="83.0" fitWidth="402.0" layoutX="903.0" layoutY="29.0" pickOnBounds="true" preserveRatio="true">
+               <image>
+                  <Image url="@../../../../../../../Desktop/Image2.png" />
+               </image>
+            </ImageView>
+            <Button fx:id="changepicbtn" layoutX="83.0" layoutY="72.0" mnemonicParsing="false" onAction="#uploadsiguppic" text="change pic" visible="false" />
+            <Button layoutX="971.0" layoutY="126.0" mnemonicParsing="false" onAction="#retourner" text="Log out" />
+         </children>
+      </AnchorPane>
+      <AnchorPane layoutX="31.0" layoutY="162.0" prefHeight="467.0" prefWidth="417.0">
+         <children>
+            <Label layoutX="20.0" layoutY="64.0" text="Name" />
+            <TextField fx:id="TFname" disable="true" layoutX="20.0" layoutY="91.0" prefHeight="31.0" prefWidth="139.0" />
+            <Label layoutX="172.0" layoutY="64.0" text="Lastname" />
+            <TextField fx:id="TFlastname" disable="true" layoutX="172.0" layoutY="91.0" prefHeight="31.0" prefWidth="139.0" />
+            <Label layoutX="22.0" layoutY="147.0" text="Email" />
+            <TextField fx:id="TFemail" disable="true" layoutX="20.0" layoutY="174.0" prefHeight="31.0" prefWidth="295.0" />
+            <Label layoutX="21.0" layoutY="223.0" text="Password" />
+            <PasswordField fx:id="TFpassword" disable="true" layoutX="20.0" layoutY="252.0" prefHeight="31.0" prefWidth="295.0" />
+            <TextField fx:id="TFrole" disable="true" layoutX="20.0" layoutY="328.0" prefHeight="31.0" prefWidth="295.0" />
+            <Label layoutX="22.0" layoutY="304.0" text="Role" />
+            <Button fx:id="modifbtn" layoutX="119.0" layoutY="407.0" mnemonicParsing="false" onAction="#modifClicked" prefHeight="31.0" prefWidth="139.0" text="Modifier" />
+         </children>
+      </AnchorPane>
+      <Button fx:id="suppbtn" layoutX="822.0" layoutY="574.0" mnemonicParsing="false" onAction="#deleteClicked" prefHeight="31.0" prefWidth="213.0" text="Supprimer mon compte" />
+      <TextField fx:id="cityInput" layoutX="769.0" layoutY="220.0" />
+      <Text fx:id="weathertext" layoutX="762.0" layoutY="321.0" strokeType="OUTSIDE" strokeWidth="0.0" wrappingWidth="213.00000903010368" AnchorPane.leftAnchor="762.0" />
+      <Button layoutX="787.0" layoutY="396.0" mnemonicParsing="false" onAction="#loadweatherdata" prefHeight="31.0" prefWidth="152.0" style="-fx-background-color: FFB823;" text="Load Weather Data " />
+      <Label layoutX="598.0" layoutY="221.0" text="Choose Your City">
+         <font>
+            <Font size="20.0" />
+         </font>
+      </Label>
+      <ImageView fitHeight="41.0" fitWidth="52.0" layoutX="975.0" layoutY="216.0" pickOnBounds="true" preserveRatio="true">
+         <image>
+            <Image url="@../../../../../../Downloads/weather.png" />
+         </image>
+      </ImageView>
+   </children>
+</AnchorPane>
