+++ Gestion-User/views/Usersmanagment.fxml
@@ -1,0 +1,41 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.control.Label?>
+<?import javafx.scene.control.TableColumn?>
+<?import javafx.scene.control.TableView?>
+<?import javafx.scene.control.TextField?>
+<?import javafx.scene.layout.AnchorPane?>
+<?import javafx.scene.text.Font?>
+
+<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.UsersmanagmentController">
+   <children>
+      <AnchorPane layoutX="-1.0" prefHeight="152.0" prefWidth="1051.0" style="-fx-background-color: ADC47F;">
+         <children>
+            <Button layoutX="932.0" layoutY="24.0" mnemonicParsing="false" onAction="#retourner" text="Retourner" />
+            <Label layoutX="220.0" layoutY="41.0" text="User Details">
+               <font>
+                  <Font size="41.0" />
+               </font>
+            </Label>
+         </children></AnchorPane>
+      <Button layoutX="605.0" layoutY="467.0" mnemonicParsing="false" onAction="#resfreshuser" text="Refresh" />
+      <TableView fx:id="tableusers" layoutX="54.0" layoutY="231.0" prefHeight="287.0" prefWidth="534.0">
+        <columns>
+          <TableColumn fx:id="colid" prefWidth="75.0" text="id" />
+          <TableColumn fx:id="colname" prefWidth="75.0" text="name" />
+            <TableColumn fx:id="collastname" prefWidth="75.0" text="lastname" />
+            <TableColumn fx:id="colemail" prefWidth="75.0" text="Email" />
+            <TableColumn fx:id="colprofilepic" prefWidth="143.0" text="profilepic" />
+            <TableColumn fx:id="colrole" prefWidth="88.0" text="role" />
+        </columns>
+      </TableView>
+      <Button layoutX="692.0" layoutY="467.0" mnemonicParsing="false" onAction="#remove" text="Remove" />
+      <Label layoutX="26.0" layoutY="175.0" prefHeight="21.0" prefWidth="102.0" text="Search User">
+         <font>
+            <Font name="System Bold" size="15.0" />
+         </font>
+      </Label>
+      <TextField fx:id="filterdata" layoutX="128.0" layoutY="170.0" prefHeight="31.0" prefWidth="325.0" promptText="keywords.." />
+   </children>
+</AnchorPane>
