--- SprintJava2/views/InterfaceAdmin.fxml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.VBox?>
-
-<AnchorPane id="AnchorPane" maxHeight="630.0" maxWidth="1043.0" minHeight="630.0" minWidth="1043.0" prefHeight="630.0" prefWidth="1043.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.InterfaceAdminController">
-   <children>
-      <AnchorPane layoutY="-6.0" prefHeight="636.0" prefWidth="200.0" style="-fx-background-color: ADC47F;">
-         <children>
-            <Button fx:id="userpagebtn" layoutX="35.0" layoutY="130.0" mnemonicParsing="false" onAction="#userpage" prefHeight="31.0" prefWidth="116.0" text="Users" />
-            <Button fx:id="homebtn" layoutX="36.0" layoutY="199.0" mnemonicParsing="false" onAction="#homepage" prefHeight="31.0" prefWidth="116.0" text="home" />
-            <Button fx:id="campingbtn" layoutX="35.0" layoutY="270.0" mnemonicParsing="false" onAction="#campingpage" prefHeight="31.0" prefWidth="116.0" text="Campings" />
-            <Button fx:id="equipmentbtn" layoutX="35.0" layoutY="336.0" mnemonicParsing="false" onAction="#equipmentpage" prefHeight="31.0" prefWidth="116.0" text="Equipments" />
-            <Button fx:id="homebtn1" cacheShape="false" centerShape="false" focusTraversable="false" layoutX="16.0" layoutY="44.0" mnemonicParsing="false" prefHeight="31.0" prefWidth="156.0" scaleShape="false" text="Nature Cruise" />
-            <Button fx:id="profilebtn" layoutX="36.0" layoutY="406.0" mnemonicParsing="false" onAction="#profilepage" prefHeight="31.0" prefWidth="123.0" text="Profile" />
-         </children>
-      </AnchorPane>
-      <AnchorPane fx:id="adminpage" layoutX="200.0" prefHeight="106.0" prefWidth="842.0">
-         <children>
-            <Label fx:id="pagelabel" layoutX="33.0" layoutY="9.0" prefHeight="88.0" prefWidth="776.0" textFill="#c93d3d" />
-            <Button fx:id="logoutbtn" layoutX="744.0" layoutY="22.0" mnemonicParsing="false" onAction="#logout" text="log out" />
-         </children>
-      </AnchorPane>
-      <VBox fx:id="vbox" layoutX="220.0" layoutY="128.0" prefHeight="477.0" prefWidth="786.0" />
-   </children>
-</AnchorPane>
