--- SprintJava2/views/InterfaceLogin.fxml
@@ -1,69 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.PasswordField?>
-<?import javafx.scene.control.TextField?>
-<?import javafx.scene.image.Image?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.text.Font?>
-
-<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="969.0" xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="edu.SprintJava2.views.InterfaceLoginController">
-   <children>
-      <AnchorPane layoutX="-10.0" layoutY="-3.0" prefHeight="109.0" prefWidth="1053.0" style="-fx-background-color: ADC47F;">
-         <children>
-            <ImageView fitHeight="101.0" fitWidth="344.0" layoutX="440.0" layoutY="6.0" pickOnBounds="true" preserveRatio="true">
-               <image>
-                  <Image url="@../../../../../../../Desktop/Image2.png" />
-               </image>
-            </ImageView>
-         </children>
-      </AnchorPane>
-      <AnchorPane layoutX="21.0" layoutY="152.0" prefHeight="467.0" prefWidth="417.0">
-         <children>
-            <Label layoutX="20.0" layoutY="64.0" text="Name" />
-            <TextField fx:id="TFname" layoutX="20.0" layoutY="91.0" prefHeight="31.0" prefWidth="139.0" />
-            <Label layoutX="172.0" layoutY="64.0" text="Lastname" />
-            <TextField fx:id="TFlastname" layoutX="172.0" layoutY="91.0" prefHeight="31.0" prefWidth="139.0" />
-            <Label layoutX="22.0" layoutY="147.0" text="Email" />
-            <TextField fx:id="TFemail" layoutX="20.0" layoutY="174.0" prefHeight="31.0" prefWidth="295.0" />
-            <Label layoutX="21.0" layoutY="223.0" text="Password" />
-            <PasswordField fx:id="TFpassword" layoutX="20.0" layoutY="252.0" prefHeight="31.0" prefWidth="295.0" />
-            <Label layoutX="122.0" layoutY="14.0" text="Sign Up" textFill="#a6bc7c">
-               <font>
-                  <Font name="System Bold" size="27.0" />
-               </font>
-            </Label>
-            <ImageView fitHeight="54.0" fitWidth="89.0" layoutX="234.0" layoutY="4.0" pickOnBounds="true" preserveRatio="true">
-               <image>
-                  <Image url="@../../../../../../../Downloads/check.png" />
-               </image>
-            </ImageView>
-            <TextField fx:id="TFrole" layoutX="20.0" layoutY="328.0" prefHeight="31.0" prefWidth="295.0" />
-            <ImageView fx:id="sigupimage" fitHeight="68.0" fitWidth="82.0" layoutX="356.0" layoutY="325.0" pickOnBounds="true" preserveRatio="true">
-               <image>
-                  <Image url="@../images/userlogo.png" />
-               </image></ImageView>
-            <Button layoutX="345.0" layoutY="407.0" mnemonicParsing="false" onAction="#uploadsiguppic" text="upload pic" />
-            <Label layoutX="22.0" layoutY="304.0" text="Role" />
-            <Button layoutX="119.0" layoutY="407.0" mnemonicParsing="false" onAction="#ajouteruser" prefHeight="31.0" prefWidth="139.0" text="Sign Up" />
-         </children>
-      </AnchorPane>
-      <AnchorPane layoutX="585.0" layoutY="155.0" prefHeight="389.0" prefWidth="399.0">
-         <children>
-            <Label layoutX="36.0" layoutY="74.0" prefHeight="21.0" prefWidth="54.0" text="Email" />
-            <TextField fx:id="TFemail2" layoutX="36.0" layoutY="105.0" prefHeight="31.0" prefWidth="281.0" />
-            <PasswordField fx:id="TFpassword2" layoutX="38.0" layoutY="207.0" prefHeight="31.0" prefWidth="281.0" />
-            <Label layoutX="36.0" layoutY="164.0" prefHeight="21.0" prefWidth="95.0" text="Password" />
-            <Button layoutX="252.0" layoutY="265.0" mnemonicParsing="false" onAction="#validate" text="Sign In" />
-            <Label layoutX="131.0" layoutY="22.0" prefHeight="21.0" prefWidth="109.0" text="Sign In" textFill="#8eb970">
-               <font>
-                  <Font size="28.0" />
-               </font>
-            </Label>
-         </children>
-      </AnchorPane>
-      <Label layoutX="88.0" layoutY="131.0" text="You don't have an account? Sign In from here" />
-   </children>
-</AnchorPane>
