--- SprintJava2/views/Usersmanagment.fxml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-
-
-<AnchorPane id="AnchorPane" prefHeight="630.0" prefWidth="1043.0" xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8.0.171" fx:controller="edu.SprintJava2.views.UsersmanagmentController">
-   <children>
-      <AnchorPane layoutX="-1.0" prefHeight="152.0" prefWidth="1051.0" style="-fx-background-color: ADC47F;" />
-      <Label fx:id="lbusers" layoutX="60.0" layoutY="170.0" prefHeight="259.0" prefWidth="727.0" text="Label" />
-      <Button layoutX="661.0" layoutY="451.0" mnemonicParsing="false" onAction="#afficheruser" text="Afficher" />
-   </children>
-</AnchorPane>
